moves fetch services logic inside component
