+++ app/javascript/src/NewService/components/FormElements/ServiceDiscoveryListItems.jsx
@@ -1,25 +1,42 @@
 // @flow
 
-import React, {useEffect} from 'react'
+import React, {useState, useEffect} from 'react'
 import {Label, Select} from 'NewService/components/FormElements'
+import {fetchData} from 'utilities/utils'
+import {BASE_PATH} from 'NewService'
 
 import type { Option } from 'NewService/types'
 
 type Props = {
-  fetchServices: (namespace: string) => Promise<void>,
-  loading: boolean,
   projects: Option[],
-  services: Option[]
+  onError: (err: string) => void
 }
 
 const ServiceDiscoveryListItems = (props: Props) => {
-  const {fetchServices, loading, projects, services} = props
+  const { projects, onError } = props
+
+  const [services, setServices] = useState([])
+  const [loading, setLoading] = useState(true)
 
   useEffect(() => {
     if (projects && projects.length) {
       fetchServices(projects[0].metadata.name)
     }
   }, [projects])
+
+  const fetchServices = async (namespace: string) => {
+    setLoading(true)
+    setServices([])
+
+    try {
+      const { services } = await fetchData<{services: Option[]}>(`${BASE_PATH}/namespaces/${namespace}/services.json`)
+      setServices(services)
+    } catch (error) {
+      onError(error.message)
+    } finally {
+      setLoading(false)
+    }
+  }
 
   return (
     <React.Fragment>
