+++ app/javascript/src/NewService/index.jsx
@@ -4,3 +4,6 @@ export {NewServiceForm, NewServiceFormWrapper} from 'NewService/components/NewSe
 export {ServiceSourceForm} from 'NewService/components/ServiceSourceForm'
 export {ServiceDiscoveryForm} from 'NewService/components/ServiceDiscoveryForm'
 export {ServiceManualForm} from 'NewService/components/ServiceManualForm'
+
+export const BASE_PATH = '/p/admin/service_discovery'
+export const PROJECTS_PATH = `${BASE_PATH}/projects.json`
